Fix card count rendering as string of zeros
PostgreSQL COUNT() returns bigint which pg serializes as a string; summing
strings concatenates instead of adding ("0"+"0"+... = "000...0").

Backend: cast all COUNT(*)/COUNT(id) aliases to ::int in submissions,
customers, portal, analytics, and owner routes so values arrive as numbers.

Frontend: wrap card_count in parseInt() in all reduce operations across
Submissions, CustomerPortal, and Portal pages as a defensive fallback.

+++ backend/src/routes/analytics.js
@@ -157,7 +157,7 @@ router.get('/top-customers', authenticate, async (req, res) => {
     const result = await db.query(
       `SELECT
          cu.id, cu.name, cu.email,
-         COUNT(DISTINCT c.id) AS card_count,
+         COUNT(DISTINCT c.id)::int AS card_count,
          COUNT(DISTINCT c.id) FILTER (WHERE c.grade IS NOT NULL) AS graded_count,
          ROUND(AVG(CASE WHEN c.grade ~ '^[0-9]+(\\.[0-9]+)?$' THEN c.grade::numeric END), 2) AS avg_grade
        FROM customers cu
